// requestly/frontend/web.mql.todo/src/types.ts
export interface Improvement {
    _id?: string
    created?: number
    description?: string
    email?: string
    estimation?: string
    name?: string
    next_steps?: string
    notes?: string
    priority?: string
    size?: string
    target_quarter?: string
    team?: string
    ticket?: string
    votes?: number
}